backend/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.3.1
python-dotenv==1.0.1
psycopg2-binary==2.9.10
supabase==2.3.1
django-filter==23.5
django-storages==1.14.2
Pillow==10.3.0
pytest==8.0.2
pytest-django==4.8.0
pytest-cov==4.1.0
black==24.2.0
flake8==7.0.0
isort==5.13.2
PyJWT==2.9.0
djangorestframework-simplejwt==5.5.0 